refactor: déplacer stats vers features/stats/
