+++ src/features/stats/__tests__/StatsPages.flow.test.tsx
@@ -2,7 +2,7 @@
 import { describe, it, expect, vi } from 'vitest';
 import { screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import StatsPage from '@/components/StatsPage';
+import StatsPage from '@/features/stats/StatsPage';
 import { renderPage } from '@/__tests__/utils/test-utils';
 
 vi.mock('@/shared/hooks/useNavigationAdapter', () => ({
