+++ src/features/stats/game/__tests__/useGameStatsPage.test.ts
@@ -1,7 +1,7 @@
 // src/__tests__/hooks/useGameStatsPage.test.ts
 import { describe, it, expect } from 'vitest';
 import { renderHook, act } from '@testing-library/react';
-import { useGameStatsPage } from '@/hooks/useGameStatsPage';
+import { useGameStatsPage } from '@/features/stats/game/useGameStatsPage';
 import { mockGames, mockPlayers } from '@/__tests__/fixtures';
 
 describe('useGameStatsPage', () => {
